chore: Move constants file to src folder
